# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Справочник автомобилий API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/car-models": {
            "get": {
                "tags": [
                    "Car Models"
                ],
                "summary": "Get list of car models",
                "description": "Returns paginated list of car models with brand information and cars count",
                "operationId": "getCarModelsList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "brand_id",
                        "in": "query",
                        "description": "Filter by brand ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CarModelResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Server error occurred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "BrandResource": {
                "title": "Brand Resource",
                "description": "Brand resource representation",
                "properties": {
                    "id": {
                        "description": "Brand ID",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Brand name",
                        "type": "string",
                        "example": "Toyota"
                    },
                    "car_models_count": {
                        "description": "Number of car models for this brand",
                        "type": "integer",
                        "example": 5
                    },
                    "cars_count": {
                        "description": "Number of cars for this brand",
                        "type": "integer",
                        "example": 25
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    }
                },
                "type": "object"
            },
            "CarModelResource": {
                "title": "Car Model Resource",
                "description": "Car model resource representation",
                "properties": {
                    "id": {
                        "description": "Car model ID",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Car model name",
                        "type": "string",
                        "example": "Camry"
                    },
                    "brand": {
                        "$ref": "#/components/schemas/BrandResource"
                    },
                    "cars_count": {
                        "description": "Number of cars for this model",
                        "type": "integer",
                        "example": 10
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format: Bearer <token>",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "api_key": {
                "type": "apiKey",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
...
}